requirements.txt
ffmpeg-python
openai
pydub
python-dotenv
pyttsx3
